Refactored code from source.c into mem.c/mem.h
