+++ tests/unit/utils/CMakeLists.txt
@@ -9,6 +9,7 @@ set(TEST_UTILS_SOURCES
   ${SACN_SRC}/sacn/source_loss.c
   ${SACN_SRC}/sacn/mem.c
   ${SACN_SRC}/sacn/sockets.c
+  ${SACN_SRC}/sacn_mock/common.c
 )
 
 sacn_add_dynamic_test(test_utils ${TEST_UTILS_SOURCES})
